+++ src/routing/osrm_routed_wrapper.cpp
@@ -17,6 +17,7 @@ OsrmRoutedWrapper::OsrmRoutedWrapper(const std::string& profile,
                 server,
                 "table",
                 "durations",
+                "distances",
                 "route",
                 "alternatives=false&steps=false&overview=full&continue_"
                 "straight=false") {
@@ -24,8 +25,7 @@ OsrmRoutedWrapper::OsrmRoutedWrapper(const std::string& profile,
 
 std::string
 OsrmRoutedWrapper::build_query(const std::vector<Location>& locations,
-                               const std::string& service,
-                               const std::string& extra_args) const {
+                               const std::string& service) const {
   // Building query for osrm-routed
   std::string query = "GET /" + service;
 
@@ -38,8 +38,11 @@ OsrmRoutedWrapper::build_query(const std::vector<Location>& locations,
   }
   query.pop_back(); // Remove trailing ';'.
 
-  if (!extra_args.empty()) {
-    query += "?" + extra_args;
+  if (service == _route_service) {
+    query += "?" + _routing_args;
+  } else {
+    assert(service == _matrix_service);
+    query += "?annotations=duration,distance";
   }
 
   query += " HTTP/1.1\r\n";
@@ -63,9 +66,19 @@ bool OsrmRoutedWrapper::duration_value_is_null(
   return matrix_entry.IsNull();
 }
 
+bool OsrmRoutedWrapper::distance_value_is_null(
+  const rapidjson::Value& matrix_entry) const {
+  return matrix_entry.IsNull();
+}
+
 UserDuration OsrmRoutedWrapper::get_duration_value(
   const rapidjson::Value& matrix_entry) const {
-  return round_cost(matrix_entry.GetDouble());
+  return round_cost<UserDuration>(matrix_entry.GetDouble());
+}
+
+UserDistance OsrmRoutedWrapper::get_distance_value(
+  const rapidjson::Value& matrix_entry) const {
+  return round_cost<UserDistance>(matrix_entry.GetDouble());
 }
 
 double