+++ src/routing/osrm_routed_wrapper.h
@@ -17,8 +17,7 @@ namespace vroom::routing {
 class OsrmRoutedWrapper : public HttpWrapper {
 private:
   std::string build_query(const std::vector<Location>& locations,
-                          const std::string& service,
-                          const std::string& extra_args) const override;
+                          const std::string& service) const override;
 
   void check_response(const rapidjson::Document& json_result,
                       const std::string& service) const override;
@@ -26,8 +25,14 @@ private:
   bool
   duration_value_is_null(const rapidjson::Value& matrix_entry) const override;
 
+  bool
+  distance_value_is_null(const rapidjson::Value& matrix_entry) const override;
+
   UserDuration
   get_duration_value(const rapidjson::Value& matrix_entry) const override;
+
+  UserDistance
+  get_distance_value(const rapidjson::Value& matrix_entry) const override;
 
   double get_total_distance(const rapidjson::Value& result) const override;