+++ src/routing/valhalla_wrapper.cpp
@@ -23,6 +23,7 @@ ValhallaWrapper::ValhallaWrapper(const std::string& profile,
                 server,
                 "sources_to_targets",
                 "sources_to_targets",
+                "sources_to_targets",
                 "route",
                 R"("directions_type":"none")") {
 }
@@ -53,8 +54,7 @@ std::string ValhallaWrapper::get_matrix_query(
 }
 
 std::string
-ValhallaWrapper::get_route_query(const std::vector<Location>& locations,
-                                 const std::string& extra_args) const {
+ValhallaWrapper::get_route_query(const std::vector<Location>& locations) const {
   // Building matrix query for Valhalla.
   std::string query = "GET /" + _route_service + "?json={\"locations\":[";
 
@@ -66,9 +66,7 @@ ValhallaWrapper::get_route_query(const std::vector<Location>& locations,
   query.pop_back(); // Remove trailing ','.
 
   query += R"(],"costing":")" + profile + "\"";
-  if (!extra_args.empty()) {
-    query += "," + extra_args;
-  }
+  query += "," + _routing_args;
   query += "}";
 
   query += " HTTP/1.1\r\n";
@@ -80,12 +78,11 @@ ValhallaWrapper::get_route_query(const std::vector<Location>& locations,
 }
 
 std::string ValhallaWrapper::build_query(const std::vector<Location>& locations,
-                                         const std::string& service,
-                                         const std::string& extra_args) const {
+                                         const std::string& service) const {
   assert(service == _matrix_service or service == _route_service);
 
   return (service == _matrix_service) ? get_matrix_query(locations)
-                                      : get_route_query(locations, extra_args);
+                                      : get_route_query(locations);
 }
 
 void ValhallaWrapper::check_response(const rapidjson::Document& json_result,
@@ -127,10 +124,23 @@ bool ValhallaWrapper::duration_value_is_null(
   return matrix_entry["time"].IsNull();
 }
 
+bool ValhallaWrapper::distance_value_is_null(
+  const rapidjson::Value& matrix_entry) const {
+  assert(matrix_entry.HasMember("distance"));
+  return matrix_entry["distance"].IsNull();
+}
+
 UserDuration ValhallaWrapper::get_duration_value(
   const rapidjson::Value& matrix_entry) const {
   assert(matrix_entry["time"].IsUint());
   return matrix_entry["time"].GetUint();
+}
+
+UserDistance ValhallaWrapper::get_distance_value(
+  const rapidjson::Value& matrix_entry) const {
+  assert(matrix_entry["distance"].IsDouble());
+  return round_cost<UserDistance>(km_to_m *
+                                  matrix_entry["distance"].GetDouble());
 }
 
 double