+++ src/routing/valhalla_wrapper.h
@@ -18,12 +18,10 @@ class ValhallaWrapper : public HttpWrapper {
 private:
   std::string get_matrix_query(const std::vector<Location>& locations) const;
 
-  std::string get_route_query(const std::vector<Location>& locations,
-                              const std::string& extra_args = "") const;
+  std::string get_route_query(const std::vector<Location>& locations) const;
 
   std::string build_query(const std::vector<Location>& locations,
-                          const std::string& service,
-                          const std::string& extra_args) const override;
+                          const std::string& service) const override;
 
   void check_response(const rapidjson::Document& json_result,
                       const std::string& service) const override;
@@ -31,8 +29,14 @@ private:
   bool
   duration_value_is_null(const rapidjson::Value& matrix_entry) const override;
 
+  bool
+  distance_value_is_null(const rapidjson::Value& matrix_entry) const override;
+
   UserDuration
   get_duration_value(const rapidjson::Value& matrix_entry) const override;
+
+  UserDistance
+  get_distance_value(const rapidjson::Value& matrix_entry) const override;
 
   double get_total_distance(const rapidjson::Value& result) const override;